Refactor all `Text.*` into a single module
