--- elm-format-lib/src/Text/Parsec.hs
@@ -1,13 +1,0 @@
-module Text.Parsec
-  ( module Text.Parsec.Pos
-  , module Text.Parsec.Error
-  , module Text.Parsec.Prim
-  , module Text.Parsec.Char
-  , module Text.Parsec.Combinator
-  ) where
-
-import Text.Parsec.Pos
-import Text.Parsec.Error
-import Text.Parsec.Prim
-import Text.Parsec.Char
-import Text.Parsec.Combinator
