--- elm-format-lib/src/Text/ParserCombinators/Parsec/Char.hs
@@ -1,5 +1,0 @@
-module Text.ParserCombinators.Parsec.Char (lower) where
-
-import qualified Text.Parsec.Char
-
-lower = Text.Parsec.Char.lower
