--- elm-format-lib/src/Text/ParserCombinators/Parsec/Combinator.hs
@@ -1,5 +1,0 @@
-module Text.ParserCombinators.Parsec.Combinator (eof) where
-
-import qualified Text.Parsec.Combinator
-
-eof = Text.Parsec.Combinator.eof
