+++ elm-format-lib/test/Parse/HelpersTest.hs
@@ -6,7 +6,7 @@ import Test.Tasty.HUnit
 import AST.V0_16
 import Parse.Helpers (parens'')
 import Parse.IParser
-import Text.ParserCombinators.Parsec.Char (lower)
+import Parse.ParsecAdapter (lower)
 
 import Parse.TestHelpers
